apply code review suggestions: reorganize packages; use ticker
