+++ dal/daltest/db_test.go
@@ -1,18 +1,19 @@
-package dal
+package daltest
 
 import (
 	"testing"
 
-	"gotest.tools/assert"
+	"github.com/gonative-cc/relayer/dal"
+	"gotest.tools/v3/assert"
 )
 
 func TestInsertTx(t *testing.T) {
-	db := initTestDB(t)
+	db := InitTestDB(t)
 
-	tx := Tx{
+	tx := dal.Tx{
 		BtcTxID: 1,
 		RawTx:   "raw-transaction-hex",
-		Status:  StatusPending,
+		Status:  dal.StatusPending,
 	}
 
 	err := db.InsertTx(tx)
@@ -24,12 +25,12 @@ func TestInsertTx(t *testing.T) {
 }
 
 func TestGetPendingTxs(t *testing.T) {
-	db := initTestDB(t)
+	db := InitTestDB(t)
 
-	transactions := []Tx{
-		{BtcTxID: 1, RawTx: "tx1-hex", Status: StatusPending},
-		{BtcTxID: 2, RawTx: "tx2-hex", Status: StatusBroadcasted},
-		{BtcTxID: 3, RawTx: "tx3-hex", Status: StatusPending},
+	transactions := []dal.Tx{
+		{BtcTxID: 1, RawTx: "tx1-hex", Status: dal.StatusPending},
+		{BtcTxID: 2, RawTx: "tx2-hex", Status: dal.StatusBroadcasted},
+		{BtcTxID: 3, RawTx: "tx3-hex", Status: dal.StatusPending},
 	}
 	for _, tx := range transactions {
 		err := db.InsertTx(tx)
@@ -42,29 +43,31 @@ func TestGetPendingTxs(t *testing.T) {
 }
 
 func TestUpdateTxStatus(t *testing.T) {
-	db := initTestDB(t)
+	db := InitTestDB(t)
 	txID := uint64(1)
-	tx := Tx{
+	tx := dal.Tx{
 		BtcTxID: txID,
 		RawTx:   "raw-transaction-hex",
-		Status:  StatusPending,
+		Status:  dal.StatusPending,
 	}
 	err := db.InsertTx(tx)
 	assert.NilError(t, err)
 
-	err = db.UpdateTxStatus(txID, StatusBroadcasted)
+	err = db.UpdateTxStatus(txID, dal.StatusBroadcasted)
 	assert.NilError(t, err)
 
 	updatedTx, err := db.GetTx(txID)
 	assert.NilError(t, err)
-	assert.Equal(t, updatedTx.Status, StatusBroadcasted)
+	assert.Equal(t, updatedTx.Status, dal.StatusBroadcasted)
 }
 
-func initTestDB(t *testing.T) *DB {
+// InitTestDB initializes an in-memory database for testing purposes.
+func InitTestDB(t *testing.T) *dal.DB {
 	t.Helper()
-	db, err := NewDB(":memory:")
-	assert.NilError(t, err)
-	assert.NilError(t, db.InitDB())
 
+	db, err := dal.NewDB(":memory:")
+	assert.NilError(t, err)
+	err = db.InitDB()
+	assert.NilError(t, err)
 	return db
 }
